update for lucene/solr 6.2.1 (org.apache.lucene.analysis.utils.CharArrayMap is moved to org.apache.lucene.analysis.CharArrayMap on lucene source, all references are updated.)
