+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>org.apache.lucene</groupId>
     <artifactId>TurkishAnalysis</artifactId>
-    <version>6.0.0</version>
+    <version>6.2.1</version>
     <packaging>jar</packaging>
 
     <name>TurkishAnalysis</name>
