+++ src/main/java/org/apache/lucene/analysis/tr/util/PatternTableFactory.java
@@ -17,7 +17,7 @@ package org.apache.lucene.analysis.tr.util;
  * limitations under the License.
  */
 
-import org.apache.lucene.analysis.util.CharArrayMap;
+import org.apache.lucene.analysis.CharArrayMap;
 
 import java.util.Collections;
 import java.util.HashMap;
